Predicting Coronary Heart Disease (CHD) over ten years using TChol and LDL categories

Predicting Coronary Heart Disease (CHD) over ten years using TChol and LDL categories

predictchd(age, gender, TChol, LDL, HDL, SBP, DBP, diabetes, smoker)

Arguments

age
A number for age

gender
A binary variables taking 1 for men and 0 for women

TChol
A number for Total cholestrol in mg/dL

LDL
A number for LDL cholesterol in mg/dL

HDL
A number for HDL cholesterol in mg/dL

SBP
A number for SBP Blood pressure in mm Hg without regard to the use of antihypertensive medication

DBP
A number for DBP blood pressure in mm Hg without regard to the use of antihypertensive medication

diabetes
A binary variable taking 1 if the participant was under treatment with insulin or oral hypoglycemic agents, if casual blood glucose determinations exceeded 150 mg/dL at two clinic visits in the original cohort, or if fasting blood glucose exceeded 140 mg/dL at the initial examination of the Offspring Study participants

smoker
A binary variable taking 1 for person who smoked during the past 12 month and 0 therwise

Source

https://www.ahajournals.org/doi/full/10.1161/01.CIR.97.18.1837

Examples

predictchd (age = 55, gender = 1, TChol = 250, LDL = 120, HDL = 39,
            SBP = 146, DBP = 88, diabetes = 0 , smoker =1)

#> Error in predictchd(age = 55, gender = 1, TChol = 250, LDL = 120, HDL = 39,     SBP = 146, DBP = 88, diabetes = 0, smoker = 1): could not find function "predictchd"

predictchd (age = 30, gender = 0, TChol = 170, LDL = 120, HDL = 39,
            SBP = 145, DBP = 88, diabetes = 0 , smoker =1)

#> Error in predictchd(age = 30, gender = 0, TChol = 170, LDL = 120, HDL = 39,     SBP = 145, DBP = 88, diabetes = 0, smoker = 1): could not find function "predictchd"